requirements.txt
langchain
langchain-openai
langgraph
langchain-community

# OpenAI
openai
# FastAPI and Server
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart

# SSE Support
sse-starlette

# Async Support
aiohttp
asyncio

# Database
aiosqlite

# Utilities
python-dotenv
pyyaml
python-dateutil
# Logging
colorlog
# Development Tools
pytest
pytest-asyncio
flake8
isort
# Memory/Embeddings (optional, for production)
# faiss-cpu==1.7.4
# sentence-transformers==2.2.2

# Deployment
gunicorn

# Type Checking
mypy
types-pyyaml
types-python-dateutil